requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
gunicorn==21.2.0